requirements.txt
daphne
django
channels_redis
dotenv